Updated Navbar and added more content
Set navbar to be fixed at the top of the page and set to `is-dark` so that the
navbar can still be seen on top of the content of the page whilst scrolling.

Updated the index.html body to provide padding around the fixed navbar, so the
page content starts after the navbar.

Added all content sections, although some sections still need proper content.